findId랑 findPassword 하나로 합침
수정 전
1) 로그인화면 컨테이너 하단에
아이디찾기 비밀번호찾기 회원가입
2) 아이디찾기와 비번찾기 화면이 별도로 존재

수정 후
1) 로그인화면
아이디/비밀번호찾기 회원가입
2) 두 개를 하나로 통합
3) 로그인화면처럼 css 수정
