+++ frontend/src/App.js
@@ -5,8 +5,7 @@ import { AppContext } from "./context/AppContext.js";
 import FirstPage from "./screens/FirstScreen.js";
 import HomePage from "./screens/MainScreen.js";
 import LoginScreen from "./screens/login/LoginScreen.js";
-import FindId from "./screens/login/FindId.js";
-import FindPassword from "./screens/login/FindPassword.js";
+import FindIdOrPassword from "./screens/login/FindIdOrPassword.js";
 import Signup from "./screens/login/Signup.js";
 import Mypage from "./screens/MyPage.js";
 
@@ -71,8 +70,7 @@ const App = () => {
                         <Route path="/" element={<FirstPage />} />
                         <Route path="/Home" element={<HomePage />} />
                         <Route path="/login" element={<LoginScreen />} />
-                        <Route path="/find-id" element={<FindId />} />
-                        <Route path="/find-password" element={<FindPassword />} />
+                        <Route path="/find" element={<FindIdOrPassword />} />
                         <Route path="/signup" element={<Signup />} />
                         <Route path="/mypage" element={<Mypage />} />
                     </Routes>
