+++ frontend/src/css/login/FindIdOrPassword.css
@@ -1,0 +1,112 @@
+/* FindId.css */
+.findId-page {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+}
+  
+.find-id-container {
+  position: absolute;
+  top: 50%;
+  right: 5%;
+  transform: translateY(-50%);
+  width: 400px;
+  max-width: 90%;
+  margin: 5px auto;
+  padding: 20px;
+  background-color: #2c2c2c;
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  text-align: center;
+}
+  
+.find-id-container h2 {
+  margin-bottom: 20px;
+  font-size: 24px;
+  color: #f5f5f5;
+}
+  
+.input-group {
+  margin-bottom: 20px;
+  text-align: left;
+}
+  
+.input-group label {
+  display: block;
+  margin-bottom: 5px;
+  font-size: 14px;
+  color: #fff;
+}
+  
+.input-group input {
+  width: 100%;
+  padding: 10px;
+  margin-top: 10px;
+  font-size: 14px;
+  border: 1px solid #ddd;
+  border-radius: 4px;
+  box-sizing: border-box;
+}
+  
+.input-group input:focus {
+  border-color: red; /* 포커스 시 테두리 색상 */
+  outline: none; /* 기본 포커스 스타일 제거 */
+}
+  
+.find-id-button {
+  width: 100%;
+  padding: 10px;
+  font-size: 16px;
+  background-color: #007bff;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+  
+.find-id-button:hover {
+  background-color: #0056b3;
+}
+  
+.result-message {
+  margin-top: 20px;
+  font-size: 14px;
+  color: #fff;
+}
+  
+.additional-buttons {
+  margin-top: 20px;
+  display: grid;
+  justify-content: center;
+  gap: 10px;
+}
+  
+.signup-button-inFindID,
+.find-password-button,
+.login-button-inFindID {
+  padding: 10px 20px;
+  font-size: 14px;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: background-color 0.3s;
+}
+  
+.find-password-button {
+  background-color: #007bff;
+  color: white;
+}
+  
+.find-password-button:hover {
+  background-color: #0056b3;
+}
+  
+.login-button-inFindID {
+  background-color: #e50914;
+  color: white;
+}
+  
+.login-button-inFindID:hover {
+  background-color: #b20710;
+}
