--- frontend/src/css/login/FindPassword.css
@@ -1,63 +1,0 @@
-/* FindPassword.css */
-.find-password-container {
-    max-width: 400px;
-    margin: 50px auto;
-    padding: 20px;
-    border-radius: 8px;
-    text-align: center;
-    background-color: #2c2c2c;
-  }
-  
-  .find-password-container h2 {
-    margin-bottom: 20px;
-    font-size: 24px;
-    color: #fff;
-  }
-  
-  .find-password-container form {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .find-password-container label {
-    margin-bottom: 5px;
-    text-align: left;
-    font-size: 14px;
-    color: #fff;
-  }
-  
-  .find-password-container input {
-    margin-bottom: 15px;
-    padding: 10px;
-    font-size: 14px;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    width: 90%;
-    max-width: 400px; /* 최대 400px로 제한 */
-  }
-  
-  
-  .find-password-container input:focus {
-    border-color: red; /* 포커스 시 테두리 색상 */
-    outline: none; /* 기본 포커스 스타일 제거 */
-  }
-  
-  .find-password-container button {
-    padding: 10px;
-    font-size: 16px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-  }
-  
-  .find-password-container button:hover {
-    background-color: #0056b3;
-  }
-  
-  .find-password-container p {
-    margin-top: 15px;
-    font-size: 14px;
-    color: #fff;
-  }
